+++ interceptor/proxy_handlers_test.go
@@ -2,10 +2,14 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"fmt"
+	"log"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
+	"os"
 	"strconv"
 	"strings"
 	"testing"
@@ -22,6 +26,26 @@ import (
 	kedanet "github.com/kedacore/http-add-on/pkg/net"
 	"github.com/kedacore/http-add-on/pkg/util"
 )
+
+var TestTlsConfig = tls.Config{}
+
+func init() {
+	caCert, err := os.ReadFile("../localhost.crt")
+	if err != nil {
+		log.Fatalf("Error getting tests certs - make sure to run make test to generate them: %v", err)
+
+	}
+	caCertPool := x509.NewCertPool()
+	caCertPool.AppendCertsFromPEM(caCert)
+	cert, err := tls.LoadX509KeyPair("../localhost.crt", "../localhost.key")
+
+	if err != nil {
+		log.Fatalf("Error getting tests certs - make sure to run make test to generate them %v", err)
+	}
+
+	TestTlsConfig.RootCAs = caCertPool
+	TestTlsConfig.Certificates = []tls.Certificate{cert}
+}
 
 // the proxy should successfully forward a request to a running server
 func TestImmediatelySuccessfulProxy(t *testing.T) {
@@ -54,6 +78,7 @@ func TestImmediatelySuccessfulProxy(t *testing.T) {
 			waitTimeout:       timeouts.WorkloadReplicas,
 			respHeaderTimeout: timeouts.ResponseHeader,
 		},
+		&tls.Config{},
 	)
 	const path = "/testfwd"
 	res, req, err := reqAndRes(path)
@@ -63,6 +88,57 @@ func TestImmediatelySuccessfulProxy(t *testing.T) {
 		originPort,
 		"testdepl",
 		"testservice",
+	))
+	req = util.RequestWithStream(req, originURL)
+	req.Host = host
+
+	hdl.ServeHTTP(res, req)
+
+	r.Equal("false", res.Header().Get("X-KEDA-HTTP-Cold-Start"), "expected X-KEDA-HTTP-Cold-Start false")
+	r.Equal(200, res.Code, "expected response code 200")
+	r.Equal("test response", res.Body.String())
+}
+
+func TestImmediatelySuccessfulProxyTLS(t *testing.T) {
+	host := fmt.Sprintf("%s.testing", t.Name())
+	r := require.New(t)
+
+	originHdl := kedanet.NewTestHTTPHandlerWrapper(
+		http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+			w.WriteHeader(200)
+			_, err := w.Write([]byte("test response"))
+			r.NoError(err)
+		}),
+	)
+	srv, originURL, err := kedanet.StartTestServer(originHdl)
+	r.NoError(err)
+	defer srv.Close()
+	originPort, err := strconv.Atoi(originURL.Port())
+	r.NoError(err)
+
+	timeouts := defaultTimeouts()
+	dialCtxFunc := retryDialContextFunc(timeouts, timeouts.DefaultBackoff())
+	waitFunc := func(context.Context, string, string) (bool, error) {
+		return false, nil
+	}
+	hdl := newForwardingHandler(
+		logr.Discard(),
+		dialCtxFunc,
+		waitFunc,
+		forwardingConfig{
+			waitTimeout:       timeouts.WorkloadReplicas,
+			respHeaderTimeout: timeouts.ResponseHeader,
+		},
+		&TestTlsConfig,
+	)
+	const path = "/testfwd"
+	res, req, err := reqAndRes(path)
+	r.NoError(err)
+	req = util.RequestWithHTTPSO(req, targetFromURL(
+		originURL,
+		originPort,
+		"testdepl",
+		"testsvc",
 	))
 	req = util.RequestWithStream(req, originURL)
 	req.Host = host
@@ -98,6 +174,7 @@ func TestWaitFailedConnection(t *testing.T) {
 			waitTimeout:       timeouts.WorkloadReplicas,
 			respHeaderTimeout: timeouts.ResponseHeader,
 		},
+		&tls.Config{},
 	)
 	stream, err := url.Parse("http://0.0.0.0:0")
 	r.NoError(err)
@@ -112,6 +189,57 @@ func TestWaitFailedConnection(t *testing.T) {
 			ScaleTargetRef: httpv1alpha1.ScaleTargetRef{
 				Service: "nosuchdepl",
 				Port:    8081,
+			},
+			TargetPendingRequests: ptr.To[int32](1234),
+		},
+	})
+	req = util.RequestWithStream(req, stream)
+	req.Host = host
+
+	hdl.ServeHTTP(res, req)
+
+	r.Equal("false", res.Header().Get("X-KEDA-HTTP-Cold-Start"), "expected X-KEDA-HTTP-Cold-Start false")
+	r.Equal(502, res.Code, "response code was unexpected")
+}
+
+func TestWaitFailedConnectionTLS(t *testing.T) {
+	const host = "TestWaitFailedConnection.testing"
+	r := require.New(t)
+
+	timeouts := defaultTimeouts()
+	backoff := timeouts.DefaultBackoff()
+	backoff.Steps = 2
+	dialCtxFunc := retryDialContextFunc(
+		timeouts,
+		backoff,
+	)
+	waitFunc := func(context.Context, string, string) (bool, error) {
+		return false, nil
+	}
+	hdl := newForwardingHandler(
+		logr.Discard(),
+		dialCtxFunc,
+		waitFunc,
+		forwardingConfig{
+			waitTimeout:       timeouts.WorkloadReplicas,
+			respHeaderTimeout: timeouts.ResponseHeader,
+		},
+		&TestTlsConfig,
+	)
+	stream, err := url.Parse("http://0.0.0.0:0")
+	r.NoError(err)
+	const path = "/testfwd"
+	res, req, err := reqAndRes(path)
+	r.NoError(err)
+	req = util.RequestWithHTTPSO(req, &httpv1alpha1.HTTPScaledObject{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: "testns",
+		},
+		Spec: httpv1alpha1.HTTPScaledObjectSpec{
+			ScaleTargetRef: httpv1alpha1.ScaleTargetRef{
+				Deployment: "nosuchdepl",
+				Service:    "nosuchdepl",
+				Port:       8081,
 			},
 			TargetPendingRequests: ptr.To[int32](1234),
 		},
@@ -147,6 +275,7 @@ func TestTimesOutOnWaitFunc(t *testing.T) {
 			waitTimeout:       timeouts.WorkloadReplicas,
 			respHeaderTimeout: timeouts.ResponseHeader,
 		},
+		&tls.Config{},
 	)
 	stream, err := url.Parse("http://1.1.1.1")
 	r.NoError(err)
@@ -161,6 +290,79 @@ func TestTimesOutOnWaitFunc(t *testing.T) {
 			ScaleTargetRef: httpv1alpha1.ScaleTargetRef{
 				Service: "nosuchsvc",
 				Port:    9091,
+			},
+			TargetPendingRequests: ptr.To[int32](1234),
+		},
+	})
+	req = util.RequestWithStream(req, stream)
+	req.Host = noSuchHost
+
+	start := time.Now()
+	hdl.ServeHTTP(res, req)
+	elapsed := time.Since(start)
+
+	t.Logf("elapsed time was %s", elapsed)
+	// serving should take at least timeouts.DeploymentReplicas, but no more than
+	// timeouts.DeploymentReplicas*4
+	r.GreaterOrEqual(elapsed, timeouts.WorkloadReplicas)
+	r.LessOrEqual(elapsed, timeouts.WorkloadReplicas*4)
+	r.Equal(502, res.Code, "response code was unexpected")
+
+	// we will always return the X-KEDA-HTTP-Cold-Start header
+	// when we are able to forward the
+	// request to the backend but not if we have failed due
+	// to a timeout from a waitFunc or earlier in the pipeline,
+	// for example, if we cannot reach the Kubernetes control
+	// plane.
+	r.Equal("", res.Header().Get("X-KEDA-HTTP-Cold-Start"), "expected X-KEDA-HTTP-Cold-Start to be empty")
+
+	// waitFunc should have been called, even though it timed out
+	waitFuncCalled := false
+	select {
+	case <-waitFuncCalledCh:
+		waitFuncCalled = true
+	default:
+	}
+
+	r.True(waitFuncCalled, "wait function was not called")
+}
+
+func TestTimesOutOnWaitFuncTLS(t *testing.T) {
+	r := require.New(t)
+
+	timeouts := defaultTimeouts()
+	timeouts.WorkloadReplicas = 25 * time.Millisecond
+	timeouts.ResponseHeader = 25 * time.Millisecond
+	dialCtxFunc := retryDialContextFunc(timeouts, timeouts.DefaultBackoff())
+
+	waitFunc, waitFuncCalledCh, finishWaitFunc := notifyingFunc()
+	defer finishWaitFunc()
+	noSuchHost := fmt.Sprintf("%s.testing", t.Name())
+
+	hdl := newForwardingHandler(
+		logr.Discard(),
+		dialCtxFunc,
+		waitFunc,
+		forwardingConfig{
+			waitTimeout:       timeouts.WorkloadReplicas,
+			respHeaderTimeout: timeouts.ResponseHeader,
+		},
+		&TestTlsConfig,
+	)
+	stream, err := url.Parse("http://1.1.1.1")
+	r.NoError(err)
+	const path = "/testfwd"
+	res, req, err := reqAndRes(path)
+	r.NoError(err)
+	req = util.RequestWithHTTPSO(req, &httpv1alpha1.HTTPScaledObject{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: "testns",
+		},
+		Spec: httpv1alpha1.HTTPScaledObjectSpec{
+			ScaleTargetRef: httpv1alpha1.ScaleTargetRef{
+				Deployment: "nosuchdepl",
+				Service:    "nosuchsvc",
+				Port:       9091,
 			},
 			TargetPendingRequests: ptr.To[int32](1234),
 		},
@@ -228,6 +430,7 @@ func TestWaitsForWaitFunc(t *testing.T) {
 			waitTimeout:       timeouts.WorkloadReplicas,
 			respHeaderTimeout: timeouts.ResponseHeader,
 		},
+		&tls.Config{},
 	)
 	const path = "/testfwd"
 	res, req, err := reqAndRes(path)
@@ -258,6 +461,71 @@ func TestWaitsForWaitFunc(t *testing.T) {
 	r.Less(elapsed, waitDur*4)
 
 	r.Equal("true", res.Header().Get("X-KEDA-HTTP-Cold-Start"), "expected X-KEDA-HTTP-Cold-Start true")
+	r.Equal(
+		originRespCode,
+		res.Code,
+		"response code was unexpected",
+	)
+}
+
+func TestWaitsForWaitFuncTLS(t *testing.T) {
+	r := require.New(t)
+
+	timeouts := defaultTimeouts()
+	dialCtxFunc := retryDialContextFunc(timeouts, timeouts.DefaultBackoff())
+
+	waitFunc, waitFuncCalledCh, finishWaitFunc := notifyingFunc()
+	const (
+		noSuchHost     = "TestWaitsForWaitFunc.test"
+		originRespCode = 201
+	)
+	testSrv, testSrvURL, err := kedanet.StartTestServer(
+		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			w.WriteHeader(originRespCode)
+		}),
+	)
+	r.NoError(err)
+	defer testSrv.Close()
+	_, originPort, err := splitHostPort(testSrvURL.Host)
+	r.NoError(err)
+	hdl := newForwardingHandler(
+		logr.Discard(),
+		dialCtxFunc,
+		waitFunc,
+		forwardingConfig{
+			waitTimeout:       timeouts.WorkloadReplicas,
+			respHeaderTimeout: timeouts.ResponseHeader,
+		},
+		&TestTlsConfig,
+	)
+	const path = "/testfwd"
+	res, req, err := reqAndRes(path)
+	r.NoError(err)
+	req = util.RequestWithHTTPSO(req, targetFromURL(
+		testSrvURL,
+		originPort,
+		"nosuchdepl",
+		"nosuchsvc",
+	))
+	req = util.RequestWithStream(req, testSrvURL)
+	req.Host = noSuchHost
+
+	// make the wait function finish after a short duration
+	const waitDur = 100 * time.Millisecond
+	go func() {
+		time.Sleep(waitDur)
+		finishWaitFunc()
+	}()
+
+	start := time.Now()
+	hdl.ServeHTTP(res, req)
+	elapsed := time.Since(start)
+	r.NoError(waitForSignal(waitFuncCalledCh, 1*time.Second))
+
+	// should take at least waitDur, but no more than waitDur*4
+	r.GreaterOrEqual(elapsed, waitDur)
+	r.Less(elapsed, waitDur*4)
+
 	r.Equal(
 		originRespCode,
 		res.Code,
@@ -298,6 +566,7 @@ func TestWaitHeaderTimeout(t *testing.T) {
 			waitTimeout:       timeouts.WorkloadReplicas,
 			respHeaderTimeout: timeouts.ResponseHeader,
 		},
+		&tls.Config{},
 	)
 	const path = "/testfwd"
 	res, req, err := reqAndRes(path)
@@ -310,6 +579,65 @@ func TestWaitHeaderTimeout(t *testing.T) {
 			ScaleTargetRef: httpv1alpha1.ScaleTargetRef{
 				Service: "testsvc",
 				Port:    9094,
+			},
+			TargetPendingRequests: ptr.To[int32](1234),
+		},
+	})
+	req = util.RequestWithStream(req, originURL)
+	req.Host = originURL.Host
+
+	hdl.ServeHTTP(res, req)
+
+	r.Equal("false", res.Header().Get("X-KEDA-HTTP-Cold-Start"), "expected X-KEDA-HTTP-Cold-Start false")
+	r.Equal(502, res.Code, "response code was unexpected")
+	close(originHdlCh)
+}
+
+func TestWaitHeaderTimeoutTLS(t *testing.T) {
+	r := require.New(t)
+
+	// the origin will wait for this channel to receive or close before it sends any data back to the
+	// proxy
+	originHdlCh := make(chan struct{})
+	originHdl := kedanet.NewTestHTTPHandlerWrapper(
+		http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+			<-originHdlCh
+			w.WriteHeader(200)
+			_, err := w.Write([]byte("test response"))
+			r.NoError(err)
+		}),
+	)
+	srv, originURL, err := kedanet.StartTestServer(originHdl)
+	r.NoError(err)
+	defer srv.Close()
+
+	timeouts := defaultTimeouts()
+	dialCtxFunc := retryDialContextFunc(timeouts, timeouts.DefaultBackoff())
+	waitFunc := func(context.Context, string, string) (bool, error) {
+		return false, nil
+	}
+	hdl := newForwardingHandler(
+		logr.Discard(),
+		dialCtxFunc,
+		waitFunc,
+		forwardingConfig{
+			waitTimeout:       timeouts.WorkloadReplicas,
+			respHeaderTimeout: timeouts.ResponseHeader,
+		},
+		&TestTlsConfig,
+	)
+	const path = "/testfwd"
+	res, req, err := reqAndRes(path)
+	r.NoError(err)
+	req = util.RequestWithHTTPSO(req, &httpv1alpha1.HTTPScaledObject{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: "testns",
+		},
+		Spec: httpv1alpha1.HTTPScaledObjectSpec{
+			ScaleTargetRef: httpv1alpha1.ScaleTargetRef{
+				Deployment: "nosuchdepl",
+				Service:    "testsvc",
+				Port:       9094,
 			},
 			TargetPendingRequests: ptr.To[int32](1234),
 		},